De-centralized library name symbols
