+++ libs/home-assistant/src/config.ts
@@ -5,3 +5,4 @@ export const CRASH_REQUESTS_PER_SEC = 'CRASH_REQUESTS_PER_SEC';
 export const WEBSOCKET_URL = 'WEBSOCKET_URL';
 export const RENDER_TIMEOUT = 'RENDER_TIMEOUT';
 export const RETRY_INTERVAL = 'RETRY_INTERVAL';
+export const LIB_HOME_ASSISTANT = Symbol('home-assistant');
