+++ libs/mqtt/src/config.ts
@@ -1,2 +1,3 @@
 export const MQTT_HOST = 'MQTT_HOST';
 export const MQTT_PORT = 'MQTT_PORT';
+export const LIB_MQTT = Symbol('mqtt');
