+++ libs/mqtt/src/contracts/index.ts
@@ -1,2 +1,1 @@
 export * from './mqtt';
-export const LIB_MQTT = Symbol('mqtt');
