+++ libs/persistence/src/config.ts
@@ -3,3 +3,4 @@ export const CIPER_KEY_SIZE = 'CIPER_KEY_SIZE';
 export const CIPHER_SECRET = 'CIPHER_SECRET';
 export const SALT_START_SIZE = 'SALT_START_SIZE';
 export const SALT_END_SIZE = 'SALT_END_SIZE';
+export const LIB_PERSISTENCE = Symbol('persistence');
