+++ libs/rules-engine/src/config.ts
@@ -1,0 +1,1 @@
+export const LIB_RULES_ENGINE = Symbol('rules-engine');
