+++ libs/rules-engine/src/modules/rules-engine.module.ts
@@ -1,0 +1,10 @@
+import { LibraryModule } from '@ccontour/utilities';
+
+import { LIB_RULES_ENGINE } from '../config';
+
+@LibraryModule({
+  library: LIB_RULES_ENGINE,
+})
+export class RulesEngineModule {
+  //
+}
