+++ libs/rules-engine/src/services/index.ts
@@ -1,0 +1,1 @@
+export * from './rules-engine.service';
