+++ libs/rules-engine/src/services/rules-engine.service.ts
@@ -1,0 +1,7 @@
+import { AutoLogService } from '@ccontour/utilities';
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class RulesEngineService {
+  constructor(private readonly logger: AutoLogService) {}
+}
