+++ libs/tty/src/config.ts
@@ -6,3 +6,4 @@ export const BLOCK_PRINT_BG = 'BLOCK_PRINT_BG';
 export const BLOCK_PRINT_FG = 'BLOCK_PRINT_FG';
 export const DISABLE_CLEAR = 'DISABLE_CLEAR';
 export const PINNED_ITEMS = 'PINNED_ITEMS';
+export const LIB_TTY = Symbol('tty');
