+++ libs/utilities/src/config.ts
@@ -3,3 +3,5 @@ export const CACHE_PROVIDER = 'CACHE_PROVIDER';
 export const REDIS_HOST = 'REDIS_HOST';
 export const REDIS_PORT = 'REDIS_PORT';
 export const REDIS_DEFAULT_TTL = 'REDIS_DEFAULT_TTL';
+
+export const LIB_UTILS = Symbol('utilities');
