+++ libs/utilities/src/contracts/meta/index.ts
@@ -1,3 +1,2 @@
 export * from './config';
-export * from './library-names';
 export * from './meta';
