--- libs/utilities/src/contracts/meta/library-names.ts
@@ -1,9 +1,0 @@
-// These characters are banned from app symbol descriptions due to breaking functionality with configuration:
-// -_.
-
-export const LIB_CONTROLLER_LOGIC = Symbol('controller-logic');
-export const LIB_HOME_ASSISTANT = Symbol('home-assistant');
-export const LIB_PERSISTENCE = Symbol('persistence');
-export const LIB_TTY = Symbol('tty');
-export const LIB_SERVER = Symbol('server');
-export const LIB_UTILS = Symbol('utilities');
