+++ package.json
@@ -136,5 +136,8 @@
     "webpack-dev-server": "4.6.0",
     "webpack-node-externals": "^3.0.0",
     "yarn": "^1.22.10"
+  },
+  "resolutions": {
+    "minimist": ">=1.2.5"
   }
 }
